[ui] Make Catalog section headers clickable
