+++ js_modules/dagster-ui/packages/ui-core/src/assets/catalog/css/SelectionSectionHeader.module.css
@@ -1,0 +1,38 @@
+.header {
+  background-color: var(--color-background-light);
+  transition:
+    background-color 50ms linear,
+    box-shadow 50ms linear;
+}
+
+.header:hover {
+  background-color: var(--color-background-light-hover);
+  box-shadow:
+    inset 0 -1px 0 0 var(--color-border-default),
+    inset 0 1px 0 0 var(--color-border-default);
+}
+
+.header:has(.button:focus-visible) {
+  box-shadow:
+    inset 0 -1px 0 0 var(--color-focus-ring),
+    inset 0 1px 0 0 var(--color-focus-ring);
+}
+
+.button {
+  flex: 1;
+  height: 44px;
+  border-radius: 0;
+}
+
+.button:focus {
+  outline: none;
+}
+
+.icon {
+  transform: rotate(-90deg);
+  transition: transform 50ms linear;
+}
+
+.isOpen.icon {
+  transform: rotate(0deg);
+}
